+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Now Transpec does dynamic code analysis!
 * Support conversion of `have(n).items` matcher ([#5](https://github.com/yujinakayama/transpec/issues/5))
+* Add `-s/--skip-dynamic-analysis` option that allows to skip dynamic analysis and convert with only static analysis
 * Consider runtime type information when converting `=~` to `match_array`
 * Rename `--commit-message` option to `--generate-commit-message`